Prevent access_limit being created with auth_bypass_ids
This commit removes the code surrounding updating an AccessLimit when
auth_bypass_ids are updated. This is in favour of using the
auth_bypass_ids column on editions instead in the previous commit.

There is an additional check on whether the access_limit hash passed in
contains a user or organisation, if not then it doesn't create a access_limit
record.