Render Govspeak inside expanded links
This changes the expanded link set presenter to render Govspeak inside
the `details` hash of an expanded link using the `DetailsPresenter`.

The reason for this is so that links to a document have the same
structure as the presented document itself (albeit perhaps with a fewer
number of presented fields).

A good example of this is related to people and roles:

If you take a look at the expanded links in this Content Item for the
roles:

https://www.gov.uk/api/content/government/people/boris-johnson

We can see that only the Govspeak is available:

```
"body": [
{
"content_type": "text/govspeak",
"content": "The Prime Minister is the leader of Her Majesty's Government and is ultimately responsible for the policy and decisions of the government.\r\n\r\nAs leader of the UK government the Prime Minister also:\r\n\r\n* oversees the [operation of the Civil Service](https://www.gov.uk/government/ministers/minister-for-the-civil-service) and government agencies\r\n* appoints members of the government\r\n* is the principal government figure in the House of Commons\r\n"
}
]
```

Whereas if we go directly to the Content Item for the role itself, we see
that the Govspeak has been rendered into HTML:

"details": {
"body": "<p>The Prime Minister is the leader of Her Majesty's Government and is ultimately responsible for the policy and decisions of the government.</p>\n\n<p>As leader of the UK government the Prime Minister also:</p>\n\n<ul>\n  <li>oversees the <a href=\"https://www.gov.uk/government/ministers/minister-for-the-civil-service\">operation of the Civil Service</a> and government agencies</li>\n  <li>appoints members of the government</li>\n  <li>is the principal government figure in the House of Commons</li>\n</ul>\n",
...
}