Support reverse links with the same link type
We want reverse links for role_appointments to come from direct links of
type `person` and `role`. Currently reverse links can't be associated
with more than one type of direct link.

The reverse links with the same type were added in:
https://github.com/alphagov/publishing-api/pull/1645

The current behaviour means that we only see the `role_appointments`
links in a person content item, and not in the role. The reason for
this is through the use of `REVERSE_LINKS.key` which only returns the
key for the first occurrence of a value, and not all the keys associated
with a value. Person happens to be above role in the `REVERSE_LINKS`
hash so the Publishing API thinks that only people need to get the
reverse links for role appointments.