Support dependency resolution with the same link type
This continues from the work done in
https://github.com/alphagov/publishing-api/pull/1647 to allow the
Publishing API to support reverse links with the same link type.

Currently because of the way we do `memo[reversed] = content_ids` this
means that if the `reversed` type is the same as one previously (as is
the case for `role_appointments` from both `person` and `role`) the
previous `content_ids` get overwritten.