Tune link expansion for governments
Content can be linked to a government, which is used by frontend apps
to display when content was published by a previous government.

To do this, some fields in the details are required. These changes
tune the fields included when linking to governments to include
everything required, and not include things that don't look
useful (like analytics_identifier, locale and schema_name).