Use MANDATORY_FIELDS when configuring government link expansion
This fixes the omission of locale. I was thinking that locale wasn't
necessary, but after fighting the govuk-content-schemas for a bit, I
think it's worth including.

If in the future the government pages are translated, then some pages
may include a non-en locale representation of the government, and
being explicit about that would be useful.