Add provider state for live edition with draft
The pact tests that run in gds-api-adapters use publishing-api to set up data for the tests.

https://github.com/alphagov/gds-api-adapters/pull/975 adds a new method to
get the "live" version of content, i.e. content that has a state of either "published" or
"unpublished". To be able to test the method properly, need to be able to test the scenario where an
edition is live, but has a new draft, to be sure that the live version is being returned.