Change the RedirectHelper not to consider previous drafts
Previously it was trying to setup redirects for drafts, so I'm
guessing if you had /a as published, a draft for /b, and then updated
the draft to be /c, a redirect would be drafted for /b to /c as well
as /a to /b.

While not being optimal, and just redirecting from /a to /c, this also
broke if you moved the draft from /b to /c and then back to /b, as the
created redirect would be circular.

This commit changes the RedirectHelper to not even consider the
previous draft (if one exists), as I think the correct behaviour can
be defined in terms of the payload, and the previously published item.