Add expansion rules for roles without organisations
In https://github.com/alphagov/publishing-api/pull/1645 I added the
expansion rules for role appointments, roles and people. I was under the
impression that any subsets of those rules would be captured, but it
turns out you have to be explicit with the paths.

Previously, the rule `[:role_appointments, :role,
:ordered_parent_organisations]` meant that dependency resolution was
only triggered on roles which were tagged to an organisation (which
isn't always necessarily the case). By adding the explicit rule
`[:role_appointments, :role]` it means we also get dependency resolution
for roles not tagged to an organisation.

I've added a test case for this situation which is a bit unusual because
the role itself ends up being a dependency for the role. Although this
is something we could neaten up I think it's fine for now. The reason
this happens is because the links go `role` -> `role_appointment` ->
`role`.