Add a Rake task to fix non-en redirects
The Publishing API generates redirects automatically when the
base_path for a new draft edition of a document differs from the
previous edition. These redirect editions start off as drafts, and
then should be published when the new edition of the content is
published.

There seems to have been a bug relating to the setting of the locale
for these redirect editions, and this resulted in them not being
published as they should have been. I believe this was fixed by [1],
but the redirects created prior to that change are at least in some
cases still drafts.

1: 53e463fc51e2bd6a41bb85dbd5fcd35e769e69c3

Because the redirects are still drafts, the old edition of the content
at the old base path in the live Content Store hasn't been replaced by
a redirect, which means not only that users aren't redirected, but
they're seeing the old edition of the content.

This Rake task attempts to rectify this situation. Given a content_id
of an affected bit of content, it will search for non-en draft
redirects. For any found, it will correct the locale, and publish
them.

This is the first part of the work to fix this, this will allow fixing
the issue for Zendesk ticket 3921182. The next step is to find any
other affected content, and apply the same fix. Once this has been
done, this Rake task and any additional related code can be removed.