Persist unpublished_at for all unpublishings
It seemed a strange anomaly that we omit this field in all types of
unpublishing other than withdrawn. This caught the Publishing Workflow
team out who found out they were sending this as part of migration for
it not to be persisted.

Even though this has no effect on public facing aspects it still seems
to make more sense to store it if provided than just ignore it.