Fix S3Exporter tests following move to JSONB
There are two issues that needed fixing here:

- Firstly, the new `payload` column is now at the end of the columns in
the table, so it appears as the last column of the CSV file.

- The second more complicated issue seems to come from the fact that the
ordering of the keys with JSONB columns isn't necessarily the same order
they were inserted with. This caused problems as the ordering the tests
were generating the payloads was different to the order they were being
retrieved from the database as. The solution is this was to retrieve the
data from the database in the tests as well. There is more potential for
these tests to be refactored to be less brittle, but this should prevent
the same problem occuring in the future.