Rewrite 'represent_downstream' to fix lint issues
Previously the 'represent_downstream' tasks relied on a couple of
top-level methods, which are now raising linting issues. Since the
methods are small, it doesn't make sense to move them to a class.
Instead, this moves the method definitions into a task that acts as
a 'lowest common denominator', which the other tasks can call.

Note that 'in_batches' is an ActiveRecord-specific method that's
equivalent to 'each_slice(1000)' of a Ruby array.

The slight downside of this approach is that it involves holding more
state in memory to run some of the bigger tasks like 'all'. This seems
like a reasonable compromise for now, and we can refactor these tasks
further if they prove to be running into frequent performance issues.