Only present auth_bypass_ids of draft editions
This changes the approach used for presenting editions to the draft
content store to only share auth_bypass_ids of editions that are in a
draft state. This is a change from the approach where live editions
share auth_bypass_ids to the draft content store.

This change has been introduced following an incident where Leicester
City Council tweeted out a link to draft content with an auth_bypass_id,
even though that content was published people could still view that
content on the draft stack and were surprised. Arguably this would have
likely been resolved much quicker if that link instead took people to a
login screen and thus appeared broken for everyone.

There was previous value to keeping the auth_bypass_ids available on a
live edition as this allowed any fact checkers to continue to have
working links even after the content was published (in a scenario where
the content was published pending a review).

In hinsight of the incident I've concluded that the risks of a leaked
link situation trump the hypothetical needs of slow fact checking and
have made this change.