Remove auth_bypass_ids during publish/unpublish commands
The auth_bypass_ids field was somewhat inconsistent to the rest of the
edition model where data that wasn't needed was kept on it after
publishing. This complicates reasoning about whether auth_bypass_ids
should be sent to the content store or not.

This change makes auth_bypass_ids consistent with access limit by
removing data set for it at the point of making content live.