Update draft and live content store on publish
This makes the behaviour of this endpoint consistent with the unpublish
endpoint where the draft and live content stores are updated when
content is published.

The main reason for adding this is because auth_bypass_ids can be part
of this data and I didn't want to set up a complex mechanism to check. I
also think this should have always been set up in the past because the
publishing time changes on a publish and thus should update
dependencies.