Discard previous path reservations when changing a draft's path
When content goes through a number of different base_paths during a
drafting each one of these ends up as a reserved path that no other
publishing application can use.

The approach I've taken to resolving this is using a change of base path
in a put content as an indicator that an application is prepared to give
up a path reservation.

There are two scenarios where we won't attempt to remove a path
reservation:

1. When the new version of the draft continues to use the same path
2. When the publishing application has a live edition published with
   the base_path. The expectation here is that the user gave up on
   the work that would have led to a base path change and thus the
   reservation should remain.

This auto discarding of reservations approach is somewhat complicated
because path reservation is something that can be created implicitly
or explicitly and we have no way to tell which of these. Rather than
making the system more complicated I felt that the best approach was
to take the action of changing path as an indicator that a publishing
app is no longer interested in the path. My hunch is that, while this
may cause a rare problem, it will solve far more situations where apps
compete over trying to use a reserved path.