Delete path reservations when discarding drafts
When discarding a draft this deletes any path reservations that were
associated with the draft as long as there is not a live edition
using that base_path that was published by the same app. When there is a
live edition the expectation is that in most cases the user is
discarding a new draft of a live edition, however there are more complex
cases, such as redirects automatically created based on a URL change
that are catered for.

Path reservation is a somewhat complex system where paths can be
reserved by users explicitly (via an endpoint) or implicitly by using
the put content endpoint. This code takes the approach that discarding a
draft is an action that shows an intent not to use the path. Ideally
we'd have decided on either implicit or explicit path for
PathReservation so this could have been simpler, however that was not to
be.