Better support publishing apps changing edition locales
Previously, if an English draft existed, and a publishing app sent a
Welsh draft for the same content_id/base_path, the Publishing API
would reject this as it conflicts with the existing English draft.

I think the intent is unambiguous here, the new draft should replace
the old one. This change makes that happen.

I'm looking at this in particular, as some content in Publisher has
the wrong locale, and this change to the Publishing API will make
fixing the locale for this affected content easier.