Better facilitate changing the locale of published content
Similar to 8471eba4553564a489f074b1b87fb39534003486, this commit aims
to make it possible to take some content, and change it's
locale. Because the locale is part of the document in the Publishing
API, this it turns out was rather difficult.

With this commit, it should be possible to publish a draft, where it
has a different locale to some content which shares the content id and
base path.

This is motivated by some changes in Publisher to allow changing the
locale, but should help other publishing apps as well.