Add rake task to allow updating of country names in specialist documents
Turns out when the work was put in to allow updating of country names in
our database, we left a large blindspot in specialist publications. This
was found when it was reported that North Macedonia had not been set as
the correct name for the country for more than two years after the name
change.

This rake task allows the relevant documents to be updated when needed,
and should be done after changes made to relevant files in
govuk-content-schemas and specialist-publisher as seen in these two
pull requests:

https://github.com/alphagov/govuk-content-schemas/pull/1011
https://github.com/alphagov/specialist-publisher/pull/1714