Fix more Abort errors
This builds upon the work done recently [1] to reduce the amount of
Abort errors [2] we've been seeing for the downstream draft and live
workers.

On further investigation we've found that other orphaned content ids can
raise the same error under different circumstances (all seem to stem
from the patch_links endpoint). Below are some scenarios where this
could happen:

1) First draft document is added and removed (whitehall) from a live
   document collection. This causes an orphan content id to be
   represented downstream live but then only to fail as the orphaned
   document only resides in the draft content store.

2) First draft edition deleted (whitehall) and then the draft is removed
   from the owning document collection. This leaves the draft document
   with no editions in Publishing API which raises the error for the
   draft & live downstream workers.

3) Foreign translation edition is unpublished (publisher) which sets the
   `content_store` field to nil for that edition in Publishing API. This
   results in errors for live and draft downstream workers.

For the errors I've investigated there doesn't seem to be actually any
bugs in the system as the code is working as intended but the lack of
filtering on the orphaned content ids is causing a lot of indirect
noise.

I'm sure there are more scenarios where this error is raised but for the
specific errors I've looked into then the simplest solution is to simply
check for an edition for the given content store which mitigates the
error for all of the above scenarios.

[1]: https://github.com/alphagov/publishing-api/pull/1890
[2]: https://github.com/alphagov/publishing-api/blob/master/app/workers/downstream_draft_worker.rb#L25-L27