Add temporary rake task to fix 'value_of_funding' values
The previous "Up to £100,000" value will now become "£10,001 to
£100,000", in preparation for a new category "Up to £10,000". The
department have confirmed that no current international development
funds are below £10,001, so "10001-to-100000" should be the new
value for these.

Once the Rake task has been executed, we can revert this commit
as the code will no longer be needed.

Trello: https://trello.com/c/eDFwuCFY/2356-5-changes-to-international-development-funding-finder