Add temporary Rake task to update pa-ur locales to pa-pk
The pa-ur locale was incorrectly added and is being replaced by
pa-pk. A small number of documents were published in this locale.

Republishing from the publishing application creates new
translations, but does not remove the incorrect ones from the
publishing-api or downstream.

This Rake task will update the incorrect translations so that
republishing will not create new translations, but instead correct
the existing documents and represent downstream.