Add temporary rake task to update protected food and drink data
For `wines` and `traditional terms for wine` with a reason for
protection of `UK geographical indication from before 2021` or `EU
agreement` we need to update the date, time and content for each entry.
As there are more than 2,000 entries, this needs to be done in bulk.

Once the rake task has been run in all environments, it can be removed.