Replace fork with official sidekiq-unique-jobs gem
We have maintained a fork of the gem mhenrixon/sidekiq-unique-jobs
at alphagov/sidekiq-unique-jobs.

The motivation for the fork was that we were running version 2
of Redis that is incompatible with version 5+ of the
sidekiq-unique-jobs gem.

We needed a fork rather than pinning the gem version, since
we backported an upstream fix from version 5 of the gem to our fork
of version 4.

This enabled us to stay on a version of the gem which supported
an older version of Redis, with an upstream fix.

For more details see https://github.com/alphagov/link-checker-api/commit/7627af9c0d703ee0f1958c514390dff75b3fc26f

Since we are now runing Redis version 3 in production, we no longer
need to maintain a fork of the gem.

Publishing API is the last app to use the fork rather than the
upstream gem. All other apps such as link-checker-api are now using
the upstream gem.

I've added the middleware as directed by the gem's documentation. I
have not enable the reaper as we don't know if we need it.

This will enable us to archive our fork.