Find final redirect path before presenting to Content Store
When we send the redirect payload to Content Store, we want to
specify the final redirect destination of the content, rather than
force multiple redirects on the user. This is to prevent
`ERR_TOO_MANY_REDIRECTS` browser errors which prevent users from
being able to follow some particularly long redirect chains.

Trello: https://trello.com/c/vFuzQYFB/2628-investigate-a-fix-for-long-redirect-chains-in-the-publishing-api-5