Add logging and update deprecated uniq_args method
While upgrading the sidekiq-unique-jobs gem. Some of the required
updates were not applied which caused some fairly subtle behavioural
changes.

The main way that this was exhibited was downtime
messages applied by Mainstream Publisher not being cleared
(see here for reference
https://publisher.integration.publishing.service.gov.uk/downtimes)

When a downtime is created, two jobs are scheduled. One for midnight
before the services downtime, and another 15 seconds after the downtime
ends.

When a worker picks up the first job, it makes several requests to
the PublishingAPI which persists a new edition of the document with the
downtime message. It then pushes these changes to the Content Store.

Once the downtime has ended, a worker picks up the second job on
Mainstream. This follows the journey as above, firstly pushing changes
to PublishingAPI, which in turn pushes changes down to the Content
Store.

After investigation, it became clear that the second job ran on
mainstream and hit PublishingAPI which persisted a new edition of the
document. However, the changes were not being pushed to the Content
Store.

This meant that the downtime message was not being cleared and was still
being shown on the GOV.UK page.

sidekiq-unique-jobs constructs a key is Redis based off the queue, class
of the worker and the arguments passsed in
(see here https://github.com/mhenrixon/sidekiq-unique-jobs#introduction)

Before a job is enqueued, it checks to see if this exists in Redis. If
it does, it employs you're chosen conflict strategy (reject, replace
retry etc.). In our case, we choose to reject the job as there should
only ever need to be a maximum of one job enqueued for each document for
each endpoint we hit on the Content Store.

This is because we pass in a content_id and action as an argument and
construct the payload within the perform method.

After some investigation we were able to establish that the
sidekiq-unique-gem was hitting our conflict strategy for the jobs
generated by the second set of requests made from Mainstream when it
tries to clear the downtime message. This meant that the
`DownstreamLiveWorker` and `DownstreamDraftWorker` which as responsible
for pushing changes to the Content Store were not being
enqueued.

The issue was being caused by the deprecation of the `uniq_args` option.
While upgrading to version 7 of the gem, `uniq_args` option was
deprecated in favour of the `lock_args_method`.

This meant that it defaulted to use all arguments passed in, rather than
the set of arguments we had passed in via the `uniq_args` method.

Due to this, several extra arguments were used to construct the key,
including govuk_request_id. This is an id that is assigned in Mainstream
and passed through associated operations, allowing us to easily track
a requests E2E journey.

Essentially, this meant that all enqueued jobs were uniq by default with
one notable exception. Jobs that were scheduled within the same set
of operations would not be unqiue as they share a govuk_request_id.

This is conjunction with the behaviour due to some missing configuration
made this bug quite tricky to fix.