Add a user to db seeds
In development this app will fall back to mock_gds_sso [1] authentication
strategy when it is not provided with a bearer token (which happens when
apps haven't configured one). If this is the case this app will error if
there isn't a user.

The motivation for fixing this was that I found out in
publishing-e2e-tests that there is an element of luck as to whether
publishing-api errors. If the first request is with a bearer token it
will create a user [2] and then all subsequent non bearer token requests
will succeed. However if the first request lacks the token it will
error.

[1]: https://github.com/alphagov/gds-sso/blob/b3675c23a4d83774af9d77121169788efbc44ad2/lib/gds-sso/warden_config.rb#L72-L86
[2]: https://github.com/alphagov/gds-sso/blob/b3675c23a4d83774af9d77121169788efbc44ad2/lib/gds-sso/bearer_token.rb#L58-L72