Migrate HMRC manuals from 'live' to 'beta'
HMRC manuals are rendered with a 'beta' banner due to custom logic in
Manuals Frontend, as we're migrating to Government Frontend we need to
set the phase value to persist the beta status.

There are questions around whether these pages should still be in 'beta'
but for now we need to keep the same phase.

Trello:
https://trello.com/c/sCpFwYz6/1213-migrate-hmrc-manuals-from-live-to-beta