Remove all the require "rails_helper" calls
Now that we have a single spec_helper file we no longer need to require
the rails_helper in each of the tests.

We don't need to replace this with spec_helper as this is required
automatically due to the --require spec_helper in .rspec.

In doing this I also removed the occasional extra unnecessary stuff in
file headers such as Ruby freeze-string commands and requiring of
spec_helper.