Add temporary rake task to update titles on hmrc_manual_Section items.
To simplify front-end rendering, we need to add the parent manual title
to hmrc manual section items.

This commit adds a task to update the values on existing records (can be
removed once this job is complete and all existing and new
hmrc_manual_sections will have the title already).