Only present change notes with a public timestamp
This fixes a test that fails for approximately 10% of builds:

```
  1) Message queue publishing puts the correct message on the queue
     Failure/Error: expect(errors).to eql([])

       expected: []
            got: [{:failed_attribute=>"TypeV4", :fragment=>"#/details/change_history/0/public_timestamp", :message=>"T...76f-aca8114b1247#", :schema=>#<Addressable::URI:0x16c60 URI:a4923dd9-f21e-5f89-b76f-aca8114b1247#>}]

       (compared using eql?)

       Diff:
       @@ -1,6 +1,11 @@
       -[]
       +[{:failed_attribute=>"TypeV4",
       +  :fragment=>"#/details/change_history/0/public_timestamp",
       +  :message=>
       +   "The property '#/details/change_history/0/public_timestamp' of type null did not match the following type: string in schema a4923dd9-f21e-5f89-b76f-aca8114b1247#",
       +  :schema=>
       +   #<Addressable::URI:0x16c60 URI:a4923dd9-f21e-5f89-b76f-aca8114b1247#>}]

     # ./spec/integration/message_queue_publishing_spec.rb:36:in `block in ensure_message_queue_payload_validates_against_notification_schema'
     # ./spec/integration/message_queue_publishing_spec.rb:30:in `ensure_message_queue_payload_validates_against_notification_schema'
     # ./spec/integration/message_queue_publishing_spec.rb:17:in `block (3 levels) in <top (required)>'
     # ./spec/spec_helper.rb:85:in `block (3 levels) in <top (required)>'
     # ./spec/spec_helper.rb:85:in `block (2 levels) in <top (required)>'
```

The source of this failure is that the application will create a
ChangeNote model prior to the point that the public timestamp is known,
yet the schema for notification and frontend require each change note to
have a public timestamp [1]. Therefore this presenter was producing data
that didn't conform to the schema.

By filtering out the change history notes without a public timestamp we
can make the schema valid.

There is a secondary bug that I'll be fixing in a subsequent commit that
a change note without a public timestamp in the drafting of a content
item should have a public timestamp set at the point of publish.

[1]: https://github.com/alphagov/govuk-content-schemas/blob/c05e8da27278ca007701abbc5296096ece82b6a8/dist/formats/news_article/frontend/schema.json#L313-L316