Refactor complex, cascading specs
It looks like I wrote these 5 years ago so politeness of past code
isn't needed. Looks like past me preferred the ultra minimal RSpec tests
with tons of set-up - this wasn't smart as they're a total pain to work
with, as I found out just now when I tried to add a new test.

This refactors them to be more conventional tests that have the 3 A's of
unit tests: Arrange, Act and Assert.