Set timestamp on change notes during publish
This is to fix an issue that can lead to there being change notes
without a public_timestamp which are invalid against the schema [1].

At the point of publishing we have established the public_updated_at
value for an edition. Therefore if there is a change note and it doesn't
have a public_timestamp it should use the edition's public_updated_at
value.

It's valid for an edition to have a change note without a
public_timestamp while it is a draft as that is a note that describes
the changes in that content, and the public timestamp may not be
established until the item is published.

[1]: https://github.com/alphagov/govuk-content-schemas/blob/0e748b701f657ebb41f871a8d8515d2eb3872228/formats/shared/definitions/publishing_api_base.jsonnet#L96-L116