Create a column to store ids from CMS's
The purpose of this column is to allow an Edition to be associated with
a collection of ids from an external CMS system.

It is intended as a place that can store the ids of the components of an
external system that is used to populate GOV.UK content. For example if
a CMS is used to manage a piece of GOV.UK content that CMS may have
multiple entities involved in a single page. By storing these IDs we can
query the Publishing API when an entity changes in an external system
and identify which Publishing API resources are affected by it.