Allow setting and retrieving cms_entity_ids
This allows Publishing API to accept the cms_entity_ids field (subject
to a schema change) and allows the ability to filter edition results
based on it's usage.

The purpose of the cms_entity_ids field is to cater for integration with
third party CMS' - at the time of writing we are experimenting with
Contenful. In a content model CMS that allows embedding it is likely
that multiple pieces of content combine to produce a GOV.UK content
item. This field allows the storing of these ids.

When a piece of content is edited in a CMS we would want to update the
necessary GOV.UK content. We can take the id from the CMS and check
cms_entity_ids to look-up which content is affected to determine what
needs to be updated.