stringify keys before logging
Sidekiq 6.0 will throw an error if it is sent things that can't be serialized to json safely. This includes ruby hashes with symbolised keys. This makes sure that keys are stringy before being sent to sidekiq.

The flag being set on initialization is so that it will force an error if invalid keys are sent. This is to reduce the risk of filling log files with noise about incorrect JSON serialization.