Mark all routes as 'gone', not just the base path
If an edition has more than one route (e.g. an atom feed, in addition to
the base path), we currently only mark the base path as 'gone' when
unpublishing the content item.  This is a bug, as we would expect all
routes to be unpublished in the same way.

This fixes that bug by ensuring we unpublish all routes.