Fix issue with updating Worldwide Organisation pages
Whenever a Worldwide Organisation page is updated in Whitehall, we
publish the changes to Publishing API.

Since approximately April 2021, this has been failing and an error being
logged in Whitehall.  For example:

```
GdsApi::HTTPUnprocessableEntity
URL: https://publishing-api.integration.govuk-internal.digital/v2/content/f4c3a215-7a30-11e4-a3cb-005056011aef
Response body:
{"error":{"code":422,"message":"base path=/world/organisations/uk-joint-delegation-to-nato conflicts with content_id=5f5bfa0f-7631-11e4-a3cb-005056011aef and locale=en","fields":{"base":["base path=/world/organisations/uk-joint-delegation-to-nato conflicts with content_id=5f5bfa0f-7631-11e4-a3cb-005056011aef and locale=en"]}}}
```

This was occuring because there is a draft edition for a placeholder
Corporate Information Page with the same base path as the Worldwide
Organisation.  It is not clear how these editions came to be present,
but are not needed and are blocking us from publishing updates.

We never noticed this issue previously as Worldwide Organisation pages
are rendered by Whitehall directly from its own database, therefore the
lack of updates to Publishing API were never an issue.

Users would see a 500 error in Whitehall when updating the Worldwide
Organisation, but the changes would appear on the live site.

Now that we are including content from the Worldwide Organisations on
International Delegation pages (which are rendered from Content Store),
this issue has now become apparent because the content in Content Store
for these organisations is very out of date.

This PR adds a migration which removes the draft duplicated pages.